split up npm module into separate files
