+++ src/npm/materialize.js
@@ -1,0 +1,1 @@
+export default () => console.log("materilze");
