+++ src/npm/vite.config.js
@@ -3,14 +3,22 @@ import { defineConfig } from "vite";
 
 export default defineConfig({
     build: {
-        // put output in folder named static so django can find the generated files
-        outDir: "static",
         lib: {
-            // Could also be a dictionary or array of multiple entry points
             entry: resolve(__dirname, "index.js"),
-            name: "MyLib",
-            // the proper extensions will be added
-            fileName: "my-lib",
+            name: "Moore NPM dependencies",
+        },
+        rollupOptions: {
+            input: {
+                codemirror: resolve(__dirname, "codemirror.js"),
+                materialize: resolve(__dirname, "materialize.js"),
+            },
+            output: [
+                {
+                    entryFileNames: ({ name }) => `${name}.js`,
+                    // format: "esm",
+                    dir: resolve(__dirname, "static"),
+                },
+            ],
         },
     },
 });
